requirements.txt
anyio==3.6.2
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
dnspython==2.3.0
fake-useragent==1.1.1
fastapi==0.92.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
lazy-model==0.0.5
lxml==4.9.2
motor==3.1.1
parse==1.19.0
pydantic==1.10.5
pyee==8.2.2
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
requests-html==0.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
